Moved Sacado product type header and moved Sacado tests
