+++ tests/fe/fe_values_view_get_function_from_local_dof_values_02.cc
@@ -25,7 +25,7 @@
 #include "../tests.h"
 
 #include <deal.II/base/quadrature_lib.h>
-#include <deal.II/base/sacado_product_type.h>
+#include <deal.II/differentiation/sacado_product_types.h>
 #include <deal.II/dofs/dof_handler.h>
 #include <deal.II/dofs/dof_renumbering.h>
 #include <deal.II/grid/grid_generator.h>
