+++ tests/sacado/CMakeLists.txt
@@ -1,0 +1,6 @@
+CMAKE_MINIMUM_REQUIRED(VERSION 2.8.12)
+INCLUDE(../setup_testsubproject.cmake)
+PROJECT(testsuite CXX)
+IF(DEAL_II_WITH_TRILINOS)
+  DEAL_II_PICKUP_TESTS()
+ENDIF()
