+++ tests/sacado/sacado_product_type_02.output
