+++ tests/sacado/sacado_product_type_03.output
