+++ tests/sacado/sacado_product_type_04.output
