+++ tests/sacado/sacado_product_type_05.cc
@@ -14,13 +14,21 @@
 // ---------------------------------------------------------------------
 
 
-// test scalar_product between tensors and symmetric tensors with Sacado
+// test ProductType with sacado
 
+#include <typeinfo>
 
-#include <deal.II/base/symmetric_tensor.h>
-#include <deal.II/base/sacado_product_type.h>
+#include <deal.II/base/tensor.h>
+#include <deal.II/differentiation/sacado_product_types.h>
 
 #include "../tests.h"
+
+template <typename T, typename U, typename CompareType>
+void check()
+{
+  AssertThrow (typeid(typename ProductType<T,U>::type) == typeid(CompareType),
+               ExcInternalError());
+}
 
 
 int main()
@@ -31,21 +39,29 @@ int main()
 
 
   // check product with Tensor<2,dim>
-  Tensor<2,2,double> t;
-  SymmetricTensor<2,2,SSdouble> st;
+  check<Tensor<1,2,Sdouble>,Tensor<1,2,Sdouble>,Sdouble >();
+  check<Tensor<1,2,SSdouble>,Tensor<1,2,SSdouble>,SSdouble  >();
+  check<Tensor<2,2,Sdouble>,Tensor<2,2,Sdouble>,Tensor<2,2,Sdouble> >();
+  check<Tensor<2,2,SSdouble>,Tensor<2,2,SSdouble>,Tensor<2,2,SSdouble>  >();
+
+  Tensor<2,2,SSdouble> t1;
+  Tensor<2,2,SSdouble> t2;
   SSdouble a(2,0,7.0);
   SSdouble b(2,1,3.0);
+  SSdouble c;
   a.val() = Sdouble(2,0,7.0);
   b.val() = Sdouble(2,1,3.0);
 
   for (unsigned int i=0; i<2; ++i)
     for (unsigned int j=0; j<2; ++j)
       {
-        t[i][j] = (1.+(i+1)*(j*2));
-        st[i][j] = 2.*a+i*j*b;
+        t1[i][j] = 2.*a+i*j;
+        t2[i][j] = 3.*b-i+j;
       }
+  t1 *t2;
 
-  deallog << scalar_product(t,st) << std::endl;
-  deallog << scalar_product(st,t) << std::endl;
+  t2 += a*t1;
 
+
+  deallog << "OK" << std::endl;
 }
