+++ tests/sacado/sacado_product_type_06.output
