+++ tests/sacado/sacado_product_type_07.cc
@@ -14,21 +14,13 @@
 // ---------------------------------------------------------------------
 
 
-// test ProductType with sacado
+// test scalar_product between tensors and symmetric tensors with Sacado
 
-#include <typeinfo>
 
-#include <deal.II/base/tensor.h>
-#include <deal.II/base/sacado_product_type.h>
+#include <deal.II/base/symmetric_tensor.h>
+#include <deal.II/differentiation/sacado_product_types.h>
 
 #include "../tests.h"
-
-template <typename T, typename U, typename CompareType>
-void check()
-{
-  AssertThrow (typeid(typename ProductType<T,U>::type) == typeid(CompareType),
-               ExcInternalError());
-}
 
 
 int main()
@@ -38,32 +30,22 @@ int main()
   initlog();
 
 
-  // check product with Tensor<1,dim>
-  check<Tensor<1,2,Sdouble>,Tensor<1,2,Sdouble>,Sdouble >();
-  check<Tensor<1,2,SSdouble>,Tensor<1,2,SSdouble>,SSdouble  >();
-
-  Tensor<1,2,SSdouble> t1;
-  Tensor<1,2,SSdouble> t2;
+  // check product with Tensor<2,dim>
+  Tensor<2,2,SSdouble> t;
+  SymmetricTensor<2,2,double> st;
   SSdouble a(2,0,7.0);
   SSdouble b(2,1,3.0);
-  SSdouble c;
   a.val() = Sdouble(2,0,7.0);
   b.val() = Sdouble(2,1,3.0);
 
   for (unsigned int i=0; i<2; ++i)
-    {
-      t1[i] = 2.*a+i;
-      t2[i] = 3.*b-i;
-    }
-  const Tensor<1,2,SSdouble> t3=t2;
-  t1 *t2;
+    for (unsigned int j=0; j<2; ++j)
+      {
+        t[i][j] = 2.*a+i*j*b;
+        st[i][j] = (1.+(i+1)*(j*2));
+      }
 
-  t2 += a*t1;
+  deallog << scalar_product(t,st) << std::endl;
+  deallog << scalar_product(st,t) << std::endl;
 
-  c = 0;
-  c += (a*b + 0.5*(t3*t3 + t1*t2))*0.3;
-
-
-
-  deallog << "OK" << std::endl;
 }
