+++ tests/sacado/sacado_product_type_08.output
